requirements.txt
python-dotenv==1.0.0
openai==0.28
langchain==0.0.198
pinecone-client==2.2.2
sentence-transformers==2.2.2
chromadb==0.5.3
